Break SymbolReferenceFinder into its own class.  This makes the merge with another branch much cleaner.
